requirements.txt
python-telegram-bot==10.1.0
PyYAML==3.13